Fixing tests updating README

+++ src/test/resources/1_schema.sql
@@ -1,5 +1,5 @@
 CREATE TABLE customers (
-id bigint, 
+id integer, 
 first_name varchar(20),
 last_name varchar(20),
 middle_name varchar(20),
